share/requirements.txt
# dflux_InteractiveAnalyzer - 최적화된 의존성 목록
# 성능과 안정성을 고려한 핵심 라이브러리만 사용

# 핵심 데이터 처리 및 분석
pandas>=2.0.0

# 시각화 (matplotlib 단독 사용)
matplotlib>=3.7.0

# 수치 계산 및 배열 처리
numpy>=1.24.0

# 이미지 처리 및 GUI 지원
Pillow>=10.0.0

# 메모리 모니터링 (선택사항)
psutil>=5.9.0

# 실행파일 생성 (선택사항 - 배포용)
pyinstaller>=5.13.0

# 개발 도구 (선택사항)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0